Format site x locations object

Convert a flat data.frame with site coordinates into a proper sf object that can then be use by other functions. This function assumes that the coordinates are given in WGS84 (longitude vs. latitude). The function automatically removes repeated coordinates from the input dataset.

fb_format_site_locations(
  input_data,
  site,
  longitude,
  latitude,
  crs = sf::st_crs(4326),
  na_rm = FALSE
)

Arguments

input_data
a data.frame in a long format (see example).

site
a character of length 1. Name of the column with site labels.

longitude
a character of length 1. Name of the column with longitude. The function assumes coordinates are WGS84 (EPSG:4326).

latitude
a character of length 1. Name of the column with latitude. The function assumes coordinates are WGS84 (EPSG:4326).

crs
a character of length 1 or an object of class crs. Coordinate Reference System (CRS) of the specified coordinates. The CRS should be a valid CRS in R. It can either be a character like "+proj=longlat +datum=WGS84 +no_defs" or as specified using sf::st_crs() like sf::st_crs(4326) the default value.

na_rm
a logical value. If TRUE remove sites with incomplete coordinates. Default is FALSE.

Value

An sf object with a site column specifying site coordinates.

Examples

library("funbiogeo")

filename <- system.file("extdata", "raw_trees_data.csv", 
                        package = "funbiogeo")
all_data <- read.csv2(filename)

head(all_data)

#>                 site            species longitude latitude count     sla
#> 1  PA_76856094010538 acer_pensylvanicum -78.60508 39.91707     7 25.2665
#> 2  PA_68287127010538 acer_pensylvanicum -77.54087 40.35293     2 25.2665
#> 3  PA_76855646010538 acer_pensylvanicum -78.30221 40.18796     4 25.2665
#> 4  PA_58581106010538 acer_pensylvanicum -79.28570 40.04243     2 25.2665
#> 5 PA_145019955010661 acer_pensylvanicum -77.05348 40.37183     3 25.2665
#> 6  PA_10716167010661 acer_pensylvanicum -78.30529 41.02784     1 25.2665
#>     n_mass wood_dens
#> 1 2.221025      0.44
#> 2 2.221025      0.44
#> 3 2.221025      0.44
#> 4 2.221025      0.44
#> 5 2.221025      0.44
#> 6 2.221025      0.44

site_locations <- fb_format_site_locations(all_data, "site", "longitude", 
                                             "latitude")
head(site_locations)

#> Simple feature collection with 6 features and 1 field
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: 39.91707 ymin: -79.2857 xmax: 41.02784 ymax: -77.05348
#> CRS:           EPSG:4326
#>                 site                   geometry
#> 1  PA_76856094010538 POINT (39.91707 -78.60508)
#> 2  PA_68287127010538 POINT (40.35293 -77.54087)
#> 3  PA_76855646010538 POINT (40.18796 -78.30221)
#> 4  PA_58581106010538  POINT (40.04243 -79.2857)
#> 5 PA_145019955010661 POINT (40.37183 -77.05348)
#> 6  PA_10716167010661 POINT (41.02784 -78.30529)