Tree species counts in Pennsylvania survey sites

This dataset contains trees species counts in sites sampled in Pennsylvania. This dataset shows the format of the argument site_species used in several functions of funbiogeo. Note that sites labels are specified in row names.

species_occs

Format

A matrix with 5,770 rows (sites) and 93 columns (species).

Examples

data("species_occs")
class(species_occs)

#> [1] "matrix" "array" 

dim(species_occs)

#> [1] 5770   93

species_occs[1:6, 1:6]

#>                   acer_negundo acer_nigrum acer_pensylvanicum acer_rubrum
#> MD_9557550010661             0           0                  0          14
#> MD_9565278010661             0           0                  0           7
#> NJ_9687057010661             0           0                  0           6
#> OH_19104898010661            0           0                  0           6
#> PA_10725597010661            0           0                  0           4
#> PA_10725883010661            0           0                  0          28
#>                   acer_saccharinum acer_saccharum
#> MD_9557550010661                 0              0
#> MD_9565278010661                 0              1
#> NJ_9687057010661                 0              1
#> OH_19104898010661                0              0
#> PA_10725597010661                0             20
#> PA_10725883010661                0              0