Aggregate site data along coarser grid

This function helps aggregating site data along a coarser grid. The idea is that you have any type of data at the site scale (diversity metrics, environmental data, etc.) but you would like to work on it at a coarser scale, or you want to visualize it at that scale. This function helps you do exactly that.

fb_aggregate_site_data(site_locations, site_data, agg_grid, fun = mean)

Arguments

site_locations
an sf object with all sites.

site_data
a matrix or data.frame containing values per sites to aggregate along the provided grid. Can have one or several columns (variables to aggregate). Row names must contain sites names as provided as site_locations.

agg_grid
a SpatRaster object (package terra). A raster of one single layer, that defines the grid along which to aggregate.

fun
the function used to aggregate points values when there are multiple points in one cell. Default is mean.

Value

A SpatRaster object with as many layers as columns in site_data.

Examples

library("funbiogeo")

data("site_locations")
data("site_species")

## Import grid ----
tavg <- system.file("extdata", "annual_mean_temp.tif", package = "funbiogeo")
tavg <- terra::rast(tavg)

## Rasterize 3 first species counts ----