Extract site x locations information from long format data

Convert a flat data.frame with site coordinates into a proper sf object that can then be used by other functions. This function assumes that the coordinates are given in WGS84 (longitude vs. latitude). The function automatically removes repeated coordinates from the input dataset.

fb_format_site_locations(
  data,
  site,
  longitude,
  latitude,
  crs = sf::st_crs(4326),
  na_rm = FALSE
)

Arguments

data

a data.frame in a long format (see example).

site

a character of length 1. Name of the column with site labels.

longitude

a character of length 1. Name of the column with longitude. The function assumes coordinates are WGS84 (EPSG:4326).

latitude

a character of length 1. Name of the column with latitude. The function assumes coordinates are WGS84 (EPSG:4326).

crs

a character of length 1 or an object of class crs. Coordinate Reference System (CRS) of the specified coordinates. The CRS should be a valid CRS in R. It can either be a character like "+proj=longlat +datum=WGS84 +no_defs" or as specified using sf::st_crs() like sf::st_crs(4326) the default value.

na_rm

a logical value. If TRUE remove sites with incomplete coordinates. Default is FALSE.

Value

An sf object with a site column specifying site coordinates.

Examples

library("funbiogeo")

filename <- system.file("extdata", "raw_mammals_data.csv", 
                        package = "funbiogeo")
all_data <- read.csv(filename)

head(all_data)
#>      site longitude latitude species count adult_body_mass gestation_length
#> 1  fb_100  5.771821 59.09736  sp_006     1           21.90            23.68
#> 2  fb_100  5.771821 59.09736  sp_033     1          777.95            36.98
#> 3  fb_100  5.771821 59.09736  sp_078     1           19.30            19.60
#> 4 fb_1000 20.271821 52.59736  sp_001     1       461900.76           235.00
#> 5 fb_1000 20.271821 52.59736  sp_002     1           21.11            19.89
#> 6 fb_1000 20.271821 52.59736  sp_005     1           31.60            24.50
#>   litter_size max_longevity sexual_maturity_age diet_breadth
#> 1        5.16          52.8               57.93            4
#> 2        4.64         168.0              265.80            8
#> 3        5.54          72.0               64.71            1
#> 4        1.25         324.0              668.20            1
#> 5        5.64          48.0               76.04           NA
#> 6        4.94          48.0               43.27           NA

site_locations <- fb_format_site_locations(all_data, "site", "longitude", 
                                           "latitude")
head(site_locations)
#> Simple feature collection with 6 features and 1 field
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: 52.09736 ymin: -9.728179 xmax: 59.09736 ymax: 21.27182
#> Geodetic CRS:  WGS 84
#>        site                   geometry
#> 1    fb_100  POINT (59.09736 5.771821)
#> 4   fb_1000  POINT (52.59736 20.27182)
#> 54  fb_1001  POINT (52.59736 20.77182)
#> 103 fb_1002  POINT (52.59736 21.27182)
#> 152 fb_1007 POINT (52.09736 -9.728179)
#> 174 fb_1008 POINT (52.09736 -9.228179)