Sites x Species of Western Europe Mammals

This dataset contains the presence/absence of mammals species in Western Europe. This dataset is derived from IUCN range maps downloaded at: https://www.iucnredlist.org/resources/spatial-data-download. In respect to the license, species names have been anonymized.

site_species

Format

A data.frame with 1,505 rows (sites) and 150 columns (1 column for site label and 149 for species occurrence).

Details

See https://github.com/frbcesab/eumammals for a complete description of the workflow used to create this dataset.

This dataset shows the format of the argument site_species used in several functions of funbiogeo.

Note that sites labels (site) is the first column of this dataset.

References

IUCN (2021) The IUCN Red List of Threatened Species. Version 2021-2. Downloaded on 2021/10/11.

Examples

data("site_species")
class(site_species)
#> [1] "data.frame"
dim(site_species)
#> [1] 1505  150
site_species[1:6, 1:6]
#>   site sp_001 sp_002 sp_003 sp_004 sp_005
#> 1    1      1      0      0      0      0
#> 2    2      1      0      0      0      0
#> 3    3      1      0      0      0      0
#> 4    4      1      0      0      0      0
#> 5    5      1      0      0      0      0
#> 6    6      1      0      0      0      1