Data: species x traits of Western Europe mammals

This dataset contains values for six functional traits of Western Europe mammals species. These values were extracted from the PanTHERIA database (Jones 2009), a database on World mammals traits available at: https://esapubs.org/archive/ecol/E090/184/metadata.htm. In respect to the IUCN license, species names have been anonymized.

See https://github.com/frbcesab/eumammals for a complete description of the workflow used to create this dataset.

This dataset shows the format of the argument species_traits used in several functions of funbiogeo.

Note that species names are specified in row names.

species_traits

Format

A matrix with 149 rows (species) and the following six columns:

adult_body_mass

adult body mass (in grams)

gestation_length

length of time of non-inactive fetal growth (in days)

litter_size

number of offspring born per litter per female

max_longevity

maximum adult age (in months)

sexual_maturity_age

age when individuals are first physically capable of reproducing (in days)

diet_breadth

number of dietary categories eaten

References

Jones KE et al. (2009) PanTHERIA: A species-level database of life history, ecology, and geography of extant and recently extinct mammals. Ecology, 90, 2648. DOI: 10.1890/08-1494.1.

Examples

data("species_traits")
class(species_traits)
#> [1] "data.frame"
dim(species_traits)
#> [1] 149   6
head(species_traits)
#>        adult_body_mass gestation_length litter_size max_longevity
#> sp_001       461900.76           235.00        1.25         324.0
#> sp_002           21.11            19.89        5.64          48.0
#> sp_003              NA               NA          NA            NA
#> sp_004              NA               NA          NA            NA
#> sp_005           31.60            24.50        4.94          48.0
#> sp_006           21.90            23.68        5.16          52.8
#>        sexual_maturity_age diet_breadth
#> sp_001              668.20            1
#> sp_002               76.04           NA
#> sp_003                  NA           NA
#> sp_004                  NA           NA
#> sp_005               43.27           NA
#> sp_006               57.93            4